Fix some test failures.
* tests/test_version_xml.py:
(TestVersionXML): We need to permit access to the source tree
branch to get the relevant info.

* tests/test_service.py:
(TestXmlRpcServer.test_run_bzr): We need to permit access to the
source tree branch to get the relevant info. But why this info is
in the log but not in out.data is puzzling me :-/

* tests/test_info_xml.py:
(TestInfoXml): Address some isolation issues and some apparent
changes in the exception dicts.
(TestInfoXml.test_info_standalone): development-subtree is not
mentioned anymore.